Struct tp_individual::mapa::Mapa

pub struct Mapa { /* private fields */ }

Representa a un mapa que contiene celdas con objetos

Implementations

impl Mapa

pub fn new(grilla: Vec<Vec<Celda>>) -> Mapa

Dada una grilla crea una instanc}ia de Mapa

pub fn crear_mapa(file_path: &str) -> Result<Mapa, Box<dyn Error>>

Instancia un mapa con sus elementos a partir de un archivo indicado en los argumentos

Argumentos

file_path: Referencia a str que contiene la direccion del mapa

Returns

Devuelve una instancia de Mapa o error

pub fn obtener_largo(&self) -> usize

Devuelve el largo del mapa, que al ser cuadrado coincide en x e y

Ejemplo

Lo que sea referido como celda°n, hace referencia a una instancia de celda

use tp_individual::mapa::Mapa;
use tp_individual::celda::Celda;
let celda = Celda::Vacio{ representacion: '_' };
let mapa = Mapa::new(vec![
                        vec![celda.clone(), celda.clone()],
                        vec![celda.clone(), celda.clone()]]);
assert_eq!(mapa.obtener_largo(), 2);

pub fn obtener_celda( &mut self, fila: usize, columna: usize ) -> Result<&mut Celda, Box<dyn Error>>

Obtiene una referencia mutable a una celda contenida en un mapa

Argumentos:

&mut self: Autorreferencia

fila: Valor de la fila que se quiere obtener (equivale a x)

columna: Valor de la columna que se quiere obtener (equivale a y)

Returns

Resultado que contiene una referencia mutable a la celda indicada, o error.

Ejemplo

Lo que sea referido como celda°n, hace referencia a una instancia de celda

use tp_individual::mapa::Mapa;
use tp_individual::celda::Celda;
let mut celda = Celda::Vacio{ representacion: '_' };
let mut mapa = Mapa::new(vec![
                        vec![celda.clone(), celda.clone()],
                        vec![celda.clone(), celda.clone()]]);
assert_eq!(mapa.obtener_celda(0, 0).unwrap(), &celda);

pub fn mostrar_mapa(&self)

Muestra el mapa en consola

pub fn borrar(&mut self, fila: usize, columna: usize)

Borra una celda en la posición indicada, para eso transforma lo contenido en la celda con una Celda::Vacio

Argumentos

&mut self: autorreferencia mutable

fila: Valor de la fila que se quiere obtener (equivale a x)

columna: Valor de la columna que se quiere obtener (equivale a y)

pub fn guardar_mapa(&self, file_path: &str) -> Result<()>

Guarda el mapa en la direccion indicada, si no existe el archivo lo crea

Argumentos

&self: autorreferencia

file_path: referencia a un string que contiene la direccion de guardado

Returns

Result de Ok o Error