Enum tp_individual::model::pieza::Pieza

pub enum Pieza {
    Dama(Info),
    Rey(Info),
    Torre(Info),
    Peon(Info),
    Alfil(Info),
    Caballo(Info),
}

Representa una pieza del ajedrez. Contiene la información de la misma. Define comportamiento común a todas las piezas.

Variants

Dama(Info)

Rey(Info)

Torre(Info)

Peon(Info)

Alfil(Info)

Caballo(Info)

Implementations

impl Pieza

fn distancia_manhattan(&self, casilla_1: &Casilla, casilla_2: &Casilla) -> i32

Funcion que devuelve la distancia manhattan entre dos casillas.

fn get_info(&self) -> &Info

Funcion que devuelve la información de la pieza correspondiente.

fn puede_capturar_diagonal(
&self,
casilla_1: &Casilla,
casilla_2: &Casilla
) -> bool

Funcion que devuelve true si la pieza puede capturar a otra en dirección diagonal. False en caso contrario.

fn puede_capturar_recta(&self, casilla_1: &Casilla, casilla_2: &Casilla) -> bool

Funcion que devuelve true si la pieza puede capturar a otra en dirección recta (misma fila o columna). False en caso contrario.

fn puede_capturar_l(&self, casilla_1: &Casilla, casilla_2: &Casilla) -> bool

Funcion que devuelve true si la pieza puede capturar a otra en dirección L. False en caso contrario.

fn puede_capturar_peon(
&self,
casilla_1: &Casilla,
casilla_2: &Casilla,
color_peon: &Color
) -> bool

Funcion que devuelve true si un peón puede capturar a otra pieza, es decir, si esta se encuentra en la dirección del peón (si es blanco por encima de este, si es negro por debajo) en diagonal a una distancia de 1. False en caso contrario.

pub fn puede_capturar(&self, otra: &Pieza) -> bool

Función que define el comportamiento común a todas las piezas. Si puede capturar a la otra pieza devolverá true, sino false.