Struct tp_individual::table::Table

pub struct Table<'a> {
    file_name: &'a str,
    file: File,
}

Fields
file_name: &'a strfile: File
Implementations

impl<'a> Table<'a>

pub fn new(file_name: &'a str) -> Result<Self, Error>

pub fn get_file_directory(&self) -> &'a str

pub fn get_file_name(&self) -> Result<&'a str, TPErrors<'static>>

Get the file name of the table

Example

./path/table.csv -> table

./table.csv -> table

pub fn resolve_select( &mut self, columns: Vec<String>, opt_conditions_as_str: Option<&str>, vector_sorting: Option<Vec<(String, bool)>>, ) -> Result<Vec<Vec<String>>, Error>

given the columns of table, conditions as str and a sorting method

it will return the result of the query

Example: Lets assume we have a DB with Nombre,Apellido SELECT Nombre FROM table WHERE Apellido = 'Doe' ORDER BY Nombre ASC;

The result will be a vector of vector of string (The content readed from the csv)

pub fn resolve_insert( &self, columns: Vec<String>, values: Vec<String>, ) -> Result<Vec<String>, Error>

given a columns and values as Vec of String

It returns the proper line to write in the csv

else returns a Error.

pub fn resolve_update( &mut self, columns: Vec<String>, values: Vec<String>, opt_conditions: Option<&str>, ) -> Result<(), Error>

Function that handles the resolve of the update query

Given the columns to update, the values to update, and the conditions as str

it will return the result of the query

fn extract_conditions( &self, index_columns: &[usize], splitted_line: &[&str], columns_from_query: &[String], ) -> (Vec<(String, Value)>, Vec<String>)

Helper to extract the conditions from the splitted line

Given a index of columns, the columns and the splitted line

we return a hash of conditions AND the line itself.

pub fn insert_line_to_csv(&mut self, line: String) -> Result<(), Error>

Function that handles the insert query

Given a line, we writte it on the ‘database’ (our csv file)

pub fn resolve_delete(&mut self, conditions: Option<&str>) -> Result<(), Error>

pub fn get_directory_where_file_is(&self) -> String

Returns the directory where the file is located Example: ./path/to/file.csv -> ./path/to Example: ./file.csv -> ./

pub fn replace_original_with_tempfile(&self) -> Result<(), FileErrors>

The approach in this work is to avoid reading the whole line on memory. So, we create a “temp” csv file with the output Then, at the end, switch names.

fn get_column_from_file(&self) -> Result<Vec<String>, Error>

gets the columns of the table as string